requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pymysql==1.1.0
pymongo==4.5.0
firebase-admin==6.2.0
pydantic==2.3.0
python-multipart==0.0.6
google-genai
mysql-connector-python
numpy==1.25.2
pandas==2.1.0
tqdm==4.66.1
requests==2.31.0